backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
cryptography==41.0.7
python-multipart==0.0.6
llama-cpp-python==0.2.20
huggingface-hub==0.19.4
tqdm==4.66.1
